# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Remittance {
  employeeCode: string;
  transferNumber: string;
  currency: string;
  sender: string;
  beneficiary: string;
}

export interface AdminSummary {
  /** @minimum 0 */
  recordCount: number;
  /** @nullable */
  fileName: string | null;
  /** @nullable */
  updatedAt: string | null;
}

export type RemittanceImportInputEncoding = typeof RemittanceImportInputEncoding[keyof typeof RemittanceImportInputEncoding];


export const RemittanceImportInputEncoding = {
  'utf-8': 'utf-8',
  base64: 'base64',
} as const;

export interface RemittanceImportInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  fileName: string;
  /** @minLength 1 */
  content: string;
  encoding: RemittanceImportInputEncoding;
}

export interface ImportResult {
  /** @minimum 0 */
  importedCount: number;
  fileName: string;
  updatedAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type QueryRemittanceParams = {
/**
 * رقم كود الموظف
 * @minLength 1
 * @maxLength 64
 */
employeeCode: string;
};