# lib/api-zod/src/generated/types/remittance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Remittance {
  employeeCode: string;
  transferNumber: string;
  currency: string;
  sender: string;
  beneficiary: string;
}