# lib/api-zod/src/generated/types/remittanceImportInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RemittanceImportInputEncoding } from './remittanceImportInputEncoding';

export interface RemittanceImportInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  fileName: string;
  /** @minLength 1 */
  content: string;
  encoding: RemittanceImportInputEncoding;
}